Argument to Blob constructor to raise ValueError if only one time point
To save computational time, it can be useful to skip processing if the
Blob is only detected at a single time point, and would therefore be
excluded later in the analysis. This commit adds an argument
`allow_length_one` to `Blob.__init__()` which by default is `True`
(keeping the original behaviour), but can be set to `False` to raise an
exception if the Blob is detected at only a single time point. The
exception can then be handled in calling code, so that 'zero lifetime'
blobs are skipped without any unnecessary processing.